Add a read-more-clicked signal to GsAppWidget
We want different behaviour for the read more links in the
installed list and the updates list, so a signal makes sense.
Move the implementation for the installed list to gs-main.c.

Also, hide the box containing the button and spinner on the
updates list. This was reserving 200 pixels of empty space
 at the right.