Improve the category pages
Make it so that the filter list doesn't scroll with the
content. To do so, we take the filter list out of the grid,
and instead use a simple custom container which allocates
its children with fixed relative sizes, to allocate the
filter list and the content at a ratio of 1:2.