packagekit-refine: don't override the state if already known
The flow is like this: the GsShell calls get_popular(),
hardcoded-popular replies with a list of GsApps without metadata.
For each app, datadir-filename finds the .desktop file, in this
case in a jhbuild directory; datadir-filename-local notes that
the app is not from /usr, and marks it installed. appstream
also knows about the app, and sets a package-name.
Finally, packagekit-refine sees the app does not have a package-id
(because it didn't originate from packagekit), so it searches
for one. At that point, it would set the installed status
according to the package, but that's wrong, because the app
is installed locally.

https://bugzilla.gnome.org/show_bug.cgi?id=707728

Signed-off-by: Richard Hughes <richard@hughsie.com>