Do not show available packages when installed results are available when searching
When we search using PackageKit we do not use a INSTALLED or AVAILABLE filter as
we want both kinds of search results. A side-effect of this is that we sometimes
return both kind of packages of the same application if an update is available.

Filter out these packages so they do not get resolved and shown in the UI.

Resolves: https://bugzilla.gnome.org/show_bug.cgi?id=707322