Improve the application details page
Limit the width of the text. Allan asked for 620px, but since
GTK+ label operates in terms of characters, I just made it 80
characters. Also tweak to padding to follow Allans recommendations.
Lastly, move the styling of the title to css, by adding style
classes for application-details-title, -summary and -description.

https://bugzilla.gnome.org/show_bug.cgi?id=707319