Add a plugin to query the PackageKit prepared-update file directly
This ensures the update list presented to the user is ready-to-go and means we
don't have to manually check the offline update is ready to go after the user
presses the button.

If you're not using the systemd-updates offline updates functionality you
probably want to be using the old packagekit-updates plugin instead.