Improve search page details
There focus keeping patch from earlier today introduced a segfault
when going back from app details to the search page - this is the
one case where the previous focus widget gets removed too early.

Avoid this by not repeating the search if the search term hasn't
changed. This also avoids an annoying wait.

At the same time, improve the scrolling handling in the same way
we do on other pages - keep the scrolling position when going
back to the search page, scroll to the top in all other cases.