Prepopulate category grid
Show empty tiles in the category page as well, and replace them
with the real thing when available. We use the size of the
category to know how many empty tiles to put there.

At the same time, show all non-empty subcategories. An arbitrary
minimum size does not really make sense, and we always want to
show the featured category which may be smaller than 4.