Add a search provider
This commit adds a GNOME shell search provider that shows
uninstalled applications.

Clicking on the gnome-software icon in the results list brings
up the same search in gnome-software (including installed apps),
and clicking on one of the results brings up the details page
for that app, with the back button going to the search results
in gnome-software.

https://bugzilla.gnome.org/show_bug.cgi?id=707594