Handle offline installation sensibly
This commit adds a new app state, GS_APP_STATE_QUEUED. Apps
go in this state when you try to install them while offline.
This is done in the plugin loader, who puts such apps on its
list of pending apps. In the UI, queued apps show a 'Pending'
label and a cancel button. When network becomes available, the
plugin loader installs queued apps. The list of queued apps is
written out to ~/.local/share/gnome-software/queued-installs,
and reloaded from there when gnome-software is started.

https://bugzilla.gnome.org/show_bug.cgi?id=710114