Do not use search using PackageKit by default
We shouldn't actually need to do this, as all packages need to be converted to
applications using the AppStream metadata. The only time this is useful is when
the search term matches the package description and not the application
description, keyword, or the comment in the desktop file.

In this instance, the application should probably be shipping an AppData file
and be refined by the AppStream metadata. Consider it another reason to ship an
AppData file :)

This typically speeds up searching by more that 600ms per request.