Allow an application to have multiple sources
This allows us to have an AppStream application that has more than one
<pkgname/> tags. This allows us to define meta-applications that consist of more
than one package, for instance a 'font' where each face is in a different package.

This needs a bit more work before things like install and remove work for
applications with multiple sources, but is certainly a small step in the right
direction.