Add a setting to be able to test filting applications by missing appdata
This is much easier to test as a setting rather than getting people to recompile
the application. It also gets us much closer to what the GNOME designers want
in vision, where we only show high quality applications with screenshots and
long descriptions.

We're not at the point where we can turn this on by default yet, but it's what
we're threatening to do for GNOME 3.14