GsUpdateMonitor: don't notify every hour for available updates
If the user has taken time to click "Not now" (and thus has
definitely seen the notification), don't bother him for at
least 6 hours.
The timeout is chosen in a way that the user is notified around
twice per day, so at the next daily shutdown he will remember to
run the updates.
If the user does nothing or closes the notification, a new
one is still sent an hour later.

https://bugzilla.gnome.org/show_bug.cgi?id=721467