Improve display of recently installed updates
Don't show the number of updates in the tab button. This is
misleading since the button is not depressed, so having the
number there gives the impression that there is _another_ tab
with that number of updates, when in fact the number refers
to the already installed updates that we are showing.

Also, don't show a refresh button for already installed updates.