Sort the results by kudo for the shell sort provider
Ensure results are sorted before returning results to GNOME Shell, as typically
only 3 results are being shown. If this is not done we show the results in the
order they are parsed, which for AppStream are the application overrides and
distro-specific additions.

We sort by the kudo rather than the rating for two reasons; the first being that
only a few applications have been rated by end users, and the second that we get
the kudo information 'for free' meaning we don't have to set any additional
GsPluginRefine flags which would take longer to return to the shell.