app widget: Avoid a hardcoded width request
Instead of using a hardcoded width_request for name_label, make sure it
is added to a size group in all the callers. This makes it appear nicer
in cramped spaces, such as the installed updates dialog, where 164 px
size request for the name label can look out of place.