Improve the sorting order on the Installed page
Previously, the sorting algorithm would take into account the install
timestamp, which happened to be 0 for most apps. But not all, so some
apps would seemingly randomly move to the bottom of the list after
reinstalling (and a timestamp > 0 through that).

Fix this by simply sorting by name.

Also, while at this, make sure the sorting algorithm treats upper and
lower case letters the same.