Use AsScreenshot from libappstream-glib in the main binary
There's no point duplicating this to maintain isolation when there's no one
interested in a libappstore plugin. GNOME Software already depends on lots of
functionality in libappstream-glib which any replacement would also have to
provide.

The libappstream (not to be confused with libappstream-glib) also has an
AsScreenshot class which could be used in place.