Reduce the spinner sizes to 32px
Latest GTK+ only supports spinners up to 32px; if we request a larger
size it just gets used as padding and messes up the layout.

We could however create a custom spinner (with a custom rotating image,
should be rather easy) if the designers think it's better to have a
larger spinner.