Require the description when searching installed applications
This slows down the initial search as we might need to parse AppStream metadata
(if not already loaded) so we can show the multiline description in the search
results page. In most cases there will be no speed regression.

This will only make a visible change when an application has local AppData, but
no long description in the remote AppStream metadata.