first run dialog: Dialog nonresizable + corrected margins
Top and bottom margin of the "Let's Go Shopping" button increased
by 21 pixels each. After consulting with aday the dialog box is
set as nonresizable. This makes the new margins and whole layout fixed.

The button looked bad in some locales because of too small top and bottom
margins. Actually, the margins depended on the dialog window size.
Now both the margins and the window size are constant.

https://bugzilla.gnome.org/show_bug.cgi?id=736174