Fix a race with displaying results from shell search provider
Clicking on an app in gnome-shell search results is supposed to bring up
the details page for that app, with the back button leading to the
search page, pre-populated with the same search that was done in
gnome-shell.

The code for this was all there, but it rarely worked properly because
we'd spawn two parallel async processes, one to switch to the search
page, and another one to switch to the details page. Whichever finished
last was the one that the user actually saw.

Fix this by properly chaining the async operations so that we'd always
end up on the details page.

https://bugzilla.gnome.org/show_bug.cgi?id=735798