sources dialog: Do teardown in dispose, instead of finalize
g_cancellable_cancel can lead to a callback from the async task. Doing
it in finalize is too late, since the template children teardown has
already happened at that point and we'd crash dereferencing NULL
pointers in get_sources_cb callback.

Fix this by moving the g_cancellable_cancel() call and the rest of our
teardown code to the dispose handler, so that it runs before the
template children destruction.