Fix another race with displaying shell search provider results
Another race that caused us to come up with the search view displayed,
instead of the details view.

We were getting the delayed "search-changed" signal callback after
having already switched to the details view. If the computer was fast
enough and managed to launch the details view in 150 ms, then the
"search-changed" signal handler ran after that and caused us to switch
back to the search view.

Fix this by ignoring the next "search-changed" signal.