Make sure back button always leads back to overview
When gnome-software is launched in the details view, the back button
should normally lead back to the main view.

However, this didn't work when gnome-software window had previously been
closed in a different view; when relaunched, gnome-software would then
remember the last view from the previous run and the back button would
go back there instead of the overview mode.

Fix this by making sure we reset the internal state when closing the
main window.