Fix a crash in the screenshot loader
We were crashing if the screenshot image was destroyed while we were
still waiting on the soup network request. When the callback arrived,
we'd already cleared all the private data and didn't handle that case
properly.

Fix this by explicitly cancelling the soup message in the teardown code
and handle SOUP_STATUS_CANCELLED in the callback to make sure we return
early in that case.

Should fix https://bugzilla.redhat.com/show_bug.cgi?id=1137027 which is
high up in the Fedora 21 crasher list.