Do case-insensitive searching of suitable keywords
As gnome-software only checked the length of the search *string*, searches like
'a b c' would pass the 3-char rule and would be executed by the plugin loader.

This caused a very high CPU utilisation, and to the user it looked like the UI
had crashed. Now, only search for keywords once they have been split,
case-folded and checked for appropriate size.