sources dialog: Improved way to break up a sentence for translation
This is the second attempt to handle the translation of the sentence
'%i applications and %i add-ons installed' correctly. This patch
explicitly splits it into 3 simple cases: no software installed,
only applications installed, and only add-ons installed, and the
4th complex case where some applications and some add-ons are
installed. In this last case the words 'and' and 'installed'
may depend on the number of apps, add-ons, or total.

This patch reverts some changes from the previous commit.

https://bugzilla.gnome.org/show_bug.cgi?id=726990