back button: Ensure that the back entry focus widget is valid
Instead of g_object_ref() use g_object_add_weak_pointer()
to monitor the finalization of the focus widget while still
allowing it to be finalized. This fixes a crash on the Back
button when the old focus widget cannot accept the focus
because it has been removed, for example when it is a list box
row which has been removed from the list box.

https://bugzilla.gnome.org/show_bug.cgi?id=744879