GsHidingBox introduced
GsHidingBox is a container which displays only as many children
as its limited size allows. Also implements a minimum subset of
features of GtkBox but is derived directly from GtkContainer rather
than GtkBox. Its first application is to hide the popular tiles
if there is not enough area on the screen to fit them all.

This version allocates the extra space to the children so they
expand and shrink as the box is resized.

https://bugzilla.gnome.org/show_bug.cgi?id=742211