offline updates: GtkExpander in the error message dialog
The previous patch turned out to be not so perfect because
the detailed error message can be too wide, not only too high.
This patch introduces: GtkExpander, wrapped label inside it,
and few other minor improvements.

This is not the final solution of the problem but is backward
compatible and can be backported to the old branches.

https://bugzilla.gnome.org/show_bug.cgi?id=748936