Hide updates UI on managed systems
It doesn't make sense to show the updates UI to a user
who does not have the trigger-offline-updates privilege,
so hide the Updates tab button in this case, and don't
notify the user about available updates.

If the user manages to open the updates tab anyway (e.g.
by using the --mode=updates commandline option), tell him
that updates are managed.

https://bugzilla.gnome.org/show_bug.cgi?id=749744