Don't show "Updates" overlay text on feature tiles
The "Updates" text on the feature tiles leads to believe that it's
possible to update an app by clicking on the tile. This is misleading
since we don't offer updates on the details page.

Remove the label until we get a way to update individual apps and start
offering updates on the details page.