appstream plugin: Fix crashes due to concurrent appstream store access
Use a mutex to serialize access to priv->store to make sure that we
don't have one thread modifying the store and another thread trying to
read from it at the same time. This could happen for instance when the
AsStore file monitors detect that the data has changed and reload and
the plugin is accessing the AsStore at the same time.

https://bugzilla.redhat.com/show_bug.cgi?id=1226585
https://bugzilla.redhat.com/show_bug.cgi?id=1210989
https://bugzilla.redhat.com/show_bug.cgi?id=1183772
https://bugzilla.redhat.com/show_bug.cgi?id=1177618
https://bugzilla.redhat.com/show_bug.cgi?id=1177235
https://bugzilla.redhat.com/show_bug.cgi?id=1176343
https://bugzilla.redhat.com/show_bug.cgi?id=1176298
https://bugzilla.redhat.com/show_bug.cgi?id=1164179
https://bugzilla.redhat.com/show_bug.cgi?id=1155344