overview: Improve error handling when featured apps loading fails
Move error handling after the filtering of the featured apps list, since
the filtering can reduce the list to being empty.

Also, while at this, make sure to return early in case of
G_IO_ERROR_CANCELLED; this would either mean (a) that the page is
reloading, and there's no reason to do any work, or (b) that we're in
dispose, and all the template children are already NULL and we can't do
anything else besides returning.