Refine the package origin even if the source ID is already known
Do separate checks to see if the source ID needs refining and if the
origin needs refining. Previously we would skip refining if we already
had the source ID (but no origin) - this can happen for packages created
by the packagekit plugin, for example.

This fixes a number of packages having "Source: Unknown" in the UI.