Don't clear failed offline update results
Use the same logic here as when viewing a successful update and avoid
unlinking the results file. We already have a guard in place that makes
the error message show up only once for a user; unlinking it only makes
it more difficult to access the error message at a later time through
command line tools, such as 'pkcon offline-status'.

https://bugs.freedesktop.org/show_bug.cgi?id=91030