category page: Make state keeping more robust
Previously, we would use priv->category for storing either the main
category or the subcategory, depending on the sitation. This made it
more difficult than neccessary to figure out the main category:
sometimes we would have to walk the parent tree, sometimes it would be
NULL, sometimes the parent would already have been destroyed.

Instead, simplify the code to always store the main category and rely on
the fact that we always have the current subcategory available on the
currently selected row.

This fixes a bunch of criticals and incorrect state when going back to
the category view after installing an app.