packagekit: Make sure progress gets correctly associated with an app
Make sure that we keep track which app we're installing, so that instead
of having a global progress, we can have one for each app.

This fixes a case where installing two or more apps at the same time
would lead to the second apps progress bar moving while the first app is
installed, and vice versa.