Improve keynav handling
Try to be more careful with search-as-you-type and make sure we only
move keyboard focus to the search entry once we are sure that we got new
text. Doing it that way makes sure we ignore any modifiers, such as
shift and caps lock and avoid focusing the search entry on those.