Support the mouse back button (and Alt+Left)
Some mice have a hardware back button which should act as a shortcut
for the Back button. This patch also adds Alt+Left (Alt+Right for
RTL locales) keyboard shortcut for the Back button where it has
been missing, that is in Sources and Updates dialogs. A similar
feature in gs-shell.c has been pushed as commit b1e180b.

https://bugzilla.gnome.org/show_bug.cgi?id=747227