search provider: Don't keep the app running forever
Only hold the app alive while the actual search is going on and let it
time out after that.

This shouldn't make any difference in regular use since the background
monitor is keeping the app alive as well. However, when the background
monitor is disabled, this lets gnome-software exit after a timeout,
making it easier to debug memleaks under valgrind.

https://bugzilla.gnome.org/show_bug.cgi?id=750634