Port GsUpdateMonitor to use the plugin loader
This ensures we 'prime' both PackageKit and the AppStream plugin caches and
means we can return results to the updates panel much faster.
This also makes the code much simpler as we now only have one place files are
downloaded, and only one place where the offline update is prepared.

As a bonus, it also makes the notification work for firmware updates provided
by fwupd and any future plugin like ostree.