updates page: Only show a spinner and no text while loading data
As per the discussion on #gnome-design, don't show the "Checking for
updates…" text while loading data from the disk and only show a spinner.

This makes the experience nicer when clicking on the updates available
notification and avoids questions such as "why does it check for updates
when it just told me updates are available?".