Fix installing web applications
We only want the PackageKit plugin to match up installed desktop files with
package names. If we do this for all components we end up searching for the
thing that installed /usr/share/app-info/xmls/fedora.xml which is appstream-data,
which we really don't want to remove.

Luckily, the PackageKit plugin gets confused and just returns an unhelpful
error in this case.

Fix the problem by only adding the extra metadata for desktop, AppData or
MetaInfo files.