Add the concept of provenance for applications
For this definition, an application has provenance if it originated from the
same place as the OS vendor.

I added this as a "default negative, plugin sets positive" action to allow
applications originating from different plugins to have a different set of
places. For instance, packages would have to match the OS vendor, but xdg-apps
might just have to come from an upstream repo that's been blessed by the OS in
some way.