Try to work out if an application is free software
The way we're doing this is checking that all parts of the license string are
SPDX tokens. This isn't ideal as there is a mismatch between common distros
like Fedora and SPDX, but it's what SUSE and Debian are doing (and AppStream!)
and seems to be a good distro-neutral way of doing this.