Move the PackageKit historical updates logic into the plugin
This does mean adding a few error codes, but means we can make showing the
offline updates failure dialog a lot easier. Since we already got all the
information we need in the update monitor it makes no sense to re-get this
data when we could just save one GError.

Plus, it's two less source files to worry about...