Add a --disable-packagekit configure argument
With xdg-app apps over an ostree base using PackageKit does not make sense.

WOGUE: If you're reading this, no, we're not dropping PackageKit support.
We're still planning to keep this enabled for distros like Fedora for a long
time, but this flag makes it easier to test using jhbuild and in continuous.