Use a hashed URL prefix for the screenshot local cache filename
For metadata that has not been run with 'appstream-util mirror-screenshots' all
it takes is for one upstream to choose the same basename for the URL as another
application before there is a clash, for instance 'Screenshot1.png'.

This scheme should also stop evil apps showing the wrong screenshot for other
applications.

Resolves: https://bugzilla.gnome.org/show_bug.cgi?id=761434