Use versioned user cache directories
If upstream changes the screenshot or remote icon content without changing the
URL in the AppData file then we never re-download the new content.

We also don't want to redownload this every time as well, so just invalidate
all the cached content when we change the minor version of gnome-software,
e.g. from 3-16 to 3-18.