Do not add multiple search results for the same app
This happens when the addon also matches the search and we were choosing the
addon in preference to the main application which obviously had a lower search
score. This lead to nonsensical results when typing exact application names in
the search bar.

Now the result of the search function is a bitfield we can easily 'or' them
together and get a same value which works for all search terms.