build: Fix distcheck with --disable-packagekit
During dist, the generated sources need to be generated since they're
included in gnome_software_SOURCES. However, if PackageKit is disabled,
the recipes for generating the files are commented out.

Remove the conditional around the source file recipes. There's nothing
PackageKit dependent as they just call gdbus-codegen on DBus interface
files in the source tree. All that's needed for keeping them out of the
non-PackageKit build is to have the conditional around
gnome_software_SOURCES.

https://bugzilla.gnome.org/show_bug.cgi?id=762625