Fix overwriting the search match value
By not overwriting this field on each search we would show the wrong search
ordering when performing multiple searches where the second set of results
contained a subset of results from the first.

Also, move the match value to a new uint property rather tha trying to stuff it
in as a metadata item.