Add Steam support
To show not-yet-installed games in the search results you will have to both
install steam yourself, and crucially log at least once using the steam client.

This support is experimental and not well tested, and distros may want to ship
with --disable-steam until a better IPC is available between GNOME Software
and the Steam client.