Only show one modal dialog at a time
Keep track of the currently shown modal dialog so that we can destroy it when
another dialog is shown. Bad things happen when we try to show more than one
modal window, e.g. windows become unclickable and the main UI is unresponsive.

One way this can be seen when going to the details panel, clicking on an update
description and then clicking 'About' in the application menu.

This also involved porting away from using gtk_dialog_run() which does not cope
well when doing this kind of stacking.

You'll need https://bugzilla.gnome.org/show_bug.cgi?id=762475 for gnome-shell
if you see the application window in shadow with no modal dialogs open.