Add support for app shortcut addition/removal
Some desktops where GNOME Software is used allow to have shortcuts to
apps and it's useful for users to be able to add/remove a shortcut from
the place where they are install applications. Hence, these changes
implement the support for adding or removing shortcuts from the app
details page.

The shortcut integration in the desktop is done using the plugin's
methods gs_plugin_add_shortcut/gs_plugin_remove_shortcut and the actual
shortcut buttons are only displayed if these methods are actually implemented.

Shortcuts can mean e.g. an icon in the desktop or in a favorites area.