Add a per-plugin GsApp cache for use by plugins
This is designed to be used when the read lock is held, and ensures that
plugins can only access GsApps that have been manually added.

This now means that the 'key' can be chosen to what makes sense, for instance
in the packagekit plugin the pkgname can be used as the desktop ID or AsAppKind
is not yet known.

Also add an example of how to use this API in the dummy backend. This just
makes the 3rd-party banner on the details page flash when searching for
'chiron' but it shows the general idea.