Add a plugin to auto-add some license information
For some distos we can infer that because an application has been installed
from a specific origin that it has to be free software. This means we can show
something relevant in the license field rather than declaring the application
as non-free.

It is of course better to ship an AppData file with the exact SPDX license
string but this information is not always available.

Distros will have to set the default values of the new GSettings keys, for
instance using 'debian-*-main' and 'https://www.debian.org/social_contract'

Based on a patch by Matthias Klumpp <mak@debian.org>, many thanks.