Export a set of headers to allow external plugins to be built
This isn't something that I think is a very good idea; the API and ABI inside
gnome-software is still changing and there's a huge benefit to getting plugins
upstream where they can undergo review. However, when we're shipping
gnome-software in enterprise-class distros we might want to allow customers to
ship thier own plugins to make various business-specific changes that don't
make sense upstream.

This might involve querying a custom LDAP server and changing the suggested apps
to reflect what groups the user is in, or might involve showing a whole new
class of applications that does not conform to the Linux-specific "application
is a desktop-file" paradigm.