Do not call PackageKit.GetPackages() at startup
We get the application list from AsStore in the appstream plugin, and I can't
think what using GetPackages is adding that AsStore is unaware of.

Although we might have to refine() a lot of applications manually for the app
version it does mean that we don't end up refine()-ing the AppStream-generated
GsApp *and* dropping the PackageKit-generated GsApp on the floor, completely
unused.

This wasn't such a big problem when we were subsuming applications from
multiple plugins, but now it is as the GsAppList object notifies us about each
unused duplicate.