Factor out gs_utils_show_error_dialog()
Factoring it out makes it possible to use the nicely designed error
messages that Allan / Rafal did in other places as well, not just in
GsUpdateMonitor.

This also switches distro upgrade error messages to use the common code,
fixing the error messages to actually show up instead of a non-working
expander in the current error messages.

https://bugzilla.redhat.com/show_bug.cgi?id=1335414