trivial: Do not set the state from AppStream metadata
We know what applications are installed (as they have a source-kind of APPDATA)
but it's not safe to proxy the state when removing applications.

I guess this is an appstream-glib bug where we're not resetting the state when
we move from AppData-to-AppStream but it seems overly fragile to rely on this
alone.