Do a second resolve pass for unadopted wildcard apps
If a plugin runs before appstream that adds a GsApp with MATCH_ANY_PREFIX,
appstream helpfully converts that app to something that is recognised.

This will not be adopted, so in this case we have to do a 2nd pass of the
adopt+refine through the plugins again to ensure the management plugin is set.