trivial: Remove shortcut link buttons from the install_remove size group
The shortcut link buttons have no relief so visually they are adding no
value by being in the same size group as the install/remove. Since the
size group is also ignoring hidden buttons, this means that when the
remove button is shown, the launch/remove buttons were being resized
which is not visually appealing.