Simplify the mode changing and back button handling
Don't use the GtkEntry as a data store when we can just use the existing 'data'
parameter, and we never pass a GsCategory *and* GsApp so we can just use data
for that too.

This fixes oddness when searching and showing the button box after returning
from the search results.