Make the GsAppTile an abstract class
These changes make the GsAppTile an abstract class and move its former
functionality to a new class called GsSummaryTile.

The reason for the change is that all app tiles should extend this
class, allowing for a more more abstract code which simplifies the use
of different app tiles in the future.