Set a reasonable cache age for system upgrades
Previously we would use G_MAXUINT for all system upgrade operations
which means to never refresh any metadata once we have _anything_
downloaded. This broke pretty horribly because Fedora tends to put out
new metadata every day together with deleting old rpms from mirrors.

Instead, this changes it so that for background cache operations we
download new system upgrade metadata every week, and once the user
interactively clicks "Download" we download it again if it's older than
a day.

This should ensure that we have metadata that has a chance of working
and referencing rpms that are actually available on mirrors.