Remove an app from the installed view when it's uninstalled
If an app is uninstalled from outside the installed view (i.e. in the
app's details view), then it would not be removed from the list of
installed applications. Moreover, the app row's button would change to
"Install" but clicking it would just try to remove the app again.

These changes hide the app row from the installed view if its state has
changed and it is now available or unknown (because we still want to
show it there when it is installing).

Signed-off-by: Richard Hughes <richard@hughsie.com>