Do not allow plugins to adopt wildcard applications
The idea here was that plugins create and add a new application from refine()
rather than both trying to fight over the wildcard application which will be
harmlessly filtered out by the plugin loader.

To make this less magic and to remove some boilerplate, also add a new vfunc
which plugins can use for this special action.