Fix weirdness when de-duplicating applications
This took some time to find. When we were adding the unique_id to the hash
table we were not duplicating the unique ID. If we then added some information
(for example, the branch in refine) the unique ID would be marked as invalidated
which meant that it was automatically re-generated whenever it was needed again.

In the meantime, when we got hash table GHashFunc matches we would run the
GEqualFunc on the now-free'd memory area and fail to see that the keys were
indeed equal.

(cherry picked from commit 6b53a8ab525ec5e46712a67248f9bbba22851a35)