Do not keep trying to refresh a broken flatpak remote
libflatpak has started returning G_IO_ERROR_NOT_FOUND if the remote is not
usable, so if it does this on non-interactive refresh then add the remote to
the hash list of remotes to ignore until the next explitit refresh.

This fixes the poor search performance when a broken remote is installed.