Fix permissions of the extra AppStream file
The temporary files that get the contents of the downloaded extra
AppStream files have 0600 as the default permissions, so when they
were moved into the system location by the helper script (run as root)
it would result in a file not readable by the user.
This patch gives reading permissions to the file before moving it.