Always show the full screen update spinner when re-getting the list of updates
We don't hide the apps we are aware already need updates in the refresh phase
(the bit where the packages actually get downloaded) but when we do the quick
check for downloaded updates after the refresh we always need to clear the
screen and show something. Although this should be nearly instant, in some
cases it takes a couple of seconds.

Quite why so long needs investigating.