Treat install-bundles-system-wide as advisory, rather than policy
If we can't use the system flatpak plugin due to permissions then fall back to
the user plugin and treat the GSetting as best effort, rather than system
policy.

This is implemented by just raising the priority ordering of one of the plugins
above the over and then ignoring the request if a GsApp has already been added.