Remove any addons that have the same source as the parent app
Sometimes applications install plugins with metainfo files as part of the main
application. This is a problem because we currently strip out any addons from
the AppStream metadata whilst building that have the same pkgname as the parent
application.

If we can stop doing this it will save ~30ms per installed metainfo file not in
the appstream data file as we no longer have to use PackageKit at startup to
look up the package name. This requires the opposite change in appstream-builder
but this gnome-software patch is harmless with the old builder logic.