Fix Flatpak store initialization
When GNOME Software is started without the refresh being called, e.g.
when lauching with the "--mode overview" option, the Flatpak plugin
tries to use the AppStream store without it being populated with the
available apps.

This patch populates the store during "setup", ensuring that in any way
that GNOME Software is started the Flatpak plugin should be have access
to its apps.