Update all related applications on a proxy application
A proxy application is an application that operates on the behalf of other
applications. Is is designed so that we can have a virtual application that can
update one or more applications at the same time.

A good example of the proxy app feature in practice is the "OS Updates" item
seen on PackageKit-based distros for when packages need updating that have no
matching AppStream entry.

This also updates the systemd-updates plugin to use the new plugin loader
feature which makes the code easier to understand and the update process less
magic.