packagekit: Fix refining update details
This makes the update notes correctly show up on the updates page again.

Note that this wasn't doing an array out of bounds write, even though it
looks like that at first glance: the memory allocated for the array was
always one element larger to accomodate the trailing nul and only this
got overwritten.

https://bugzilla.gnome.org/show_bug.cgi?id=775884